1-list_len.c

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -7,9 +7,10 @@
  *
  * Return: The number of elements in h.
  */
+
 size_t list_len(const list_t *h)
 {
-	size_t elements = 0;
+	unsigned int elements = 0;
 
 	while (h)
 	{
